+++ src/DotNext.IO/ExceptionMessages.cs
@@ -22,7 +22,9 @@ internal static class ExceptionMessages
 
     internal static string WriterInReadMode => (string)Resources.Get();
 
-    internal static string NoConsumerProvided => (string)Resources.Get();
-
     internal static string FileHandleClosed => (string)Resources.Get();
+
+    internal static string ReadBufferNotEmpty => (string)Resources.Get();
+
+    internal static string WriteBufferNotEmpty => (string)Resources.Get();
 }