+++ src/DotNext.Tests/Net/Http/HttpEndPointTests.cs
@@ -1,4 +1,6 @@
 using System.Net.Sockets;
+using System.Text;
+using Microsoft.AspNetCore.Components.Forms;
 
 namespace DotNext.Net.Http;
 
@@ -63,5 +65,47 @@ public sealed class HttpEndPointTests : Test
         var uri = ep.CreateUriBuilder().Uri;
         Equal(Uri.UriSchemeHttps, uri.Scheme, ignoreCase: true);
         Equal(3262, ep.Port);
+    }
+
+    [Fact]
+    public static void Format()
+    {
+        const string expected = "http://localhost:3262/";
+        Span<char> buffer = stackalloc char[64];
+        ISpanFormattable formattable = new HttpEndPoint(new Uri(expected));
+        True(formattable.TryFormat(buffer, out var charsWritten, ReadOnlySpan<char>.Empty, provider: null));
+        
+        Equal(expected, buffer.Slice(0, charsWritten));
+        Equal(expected, formattable.ToString(format: null, formatProvider: null));
+    }
+    
+    [Fact]
+    public static void FormatAsUtf8()
+    {
+        const string expected = "http://localhost:3262/";
+        var expectedBytes = Encoding.UTF8.GetBytes(expected);
+        
+        Span<byte> buffer = stackalloc byte[64];
+        IUtf8SpanFormattable formattable = new HttpEndPoint(new Uri(expected));
+        True(formattable.TryFormat(buffer, out var charsWritten, ReadOnlySpan<char>.Empty, provider: null));
+        
+        Equal(expectedBytes, buffer.Slice(0, charsWritten));
+    }
+
+    [Fact]
+    public static void Parse()
+    {
+        const string expected = "http://localhost:3262/";
+        Equal(expected, Parse<HttpEndPoint>(expected).ToString());
+        True(TryParse<HttpEndPoint>(expected, out var ep));
+        Equal(expected, ep.ToString());
+        
+        static T Parse<T>(string input)
+            where T : IParsable<T>
+            => T.Parse(input, provider: null);
+
+        static bool TryParse<T>(string input, out T result)
+            where T : IParsable<T>
+            => T.TryParse(input, provider: null, out result);
     }
 }