+++ src/DotNext.Unsafe/Buffers/UnmanagedMemory.cs
@@ -1,5 +1,6 @@
 using System.Buffers;
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -79,7 +80,7 @@ public static class UnmanagedMemory
 
         if (length > 0)
         {
-            MemoryManager<T> manager = new UnmanagedMemory<T>((nint)pointer, length);
+            var manager = new UnmanagedMemory<T>((nint)pointer, length);
 
             // GC perf: manager doesn't own the memory represented by the pointer, no need to call Dispose from finalizer
             GC.SuppressFinalize(manager);
@@ -146,6 +147,8 @@ internal unsafe class UnmanagedMemory<T> : MemoryManager<T>
     public sealed override Span<T> GetSpan()
         => address is not null ? new(address, Length) : [];
 
+    public sealed override Memory<T> Memory => address is not null ? CreateMemory(Length) : Memory<T>.Empty;
+
     public sealed override MemoryHandle Pin(int elementIndex = 0)
     {
         ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual((uint)elementIndex, (uint)Length, nameof(elementIndex));
@@ -167,6 +170,9 @@ internal unsafe class UnmanagedMemory<T> : MemoryManager<T>
         address = null;
         Length = 0;
     }
+
+    [SuppressMessage("Reliability", "CA2015", Justification = "The caller must hold the reference to the memory object.")]
+    ~UnmanagedMemory() => Dispose(disposing: false);
 }
 
 internal class UnmanagedMemoryOwner<T> : UnmanagedMemory<T>, IUnmanagedMemory<T>