+++ CHANGELOG.md
@@ -1,6 +1,42 @@
 Release Notes
 ====
 
+# 12-29-2024
+This release is aimed to improve AOT compatibility. All the examples in the repo are now AOT compatible.
+<a href="https://www.nuget.org/packages/dotnext/5.17.0">DotNext 5.17.0</a>
+* Fixed AOT compatibility in `TaskType` class
+* Added [ISpanFormattable](https://learn.microsoft.com/en-us/dotnet/api/system.ispanformattable) and [IParsable&lt;T&gt;](https://learn.microsoft.com/en-us/dotnet/api/system.iparsable-1) interfaces to `HttpEndPoint`
+* Introduced `TryEncodeAsUtf8` extension method for `SpanWriter<T>`
+* Added more factory methods to `DotNext.Buffers.Memory` class to create [ReadOnlySequence&lt;T&gt;](https://learn.microsoft.com/en-us/dotnet/api/system.buffers.readonlysequence-1)
+* `Intrinsics.KeepAlive` is introduced for value types
+* Added `Synchronization.Wait()` synchronous methods for blocking wait of [value tasks](https://learn.microsoft.com/en-us/dotnet/api/system.threading.tasks.valuetask) without wait handles
+
+<a href="https://www.nuget.org/packages/dotnext.metaprogramming/5.17.0">DotNext.Metaprogramming 5.17.0</a>
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.unsafe/5.17.0">DotNext.Unsafe 5.17.0</a>
+* Improved AOT support
+* Fixed finalizer for unmanaged memory manager that allows to release the allocated unmanaged memory automatically by GC to avoid memory leak
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.threading/5.17.0">DotNext.Threading 5.17.0</a>
+* Improved AOT support
+
+<a href="https://www.nuget.org/packages/dotnext.io/5.17.0">DotNext.IO 5.17.0</a>
+* Reduced memory consumption for applications that use `FileReader` and `FileWriter` classes. These classes are now implemented by using lazy buffer pattern. It means that the different instances can reuse the same buffer taken from the pool
+* Fixed [255](https://github.com/dotnet/dotNext/issues/255)
+* `PoolingBufferedStream` is introduced to replace classic [BufferedStream](https://learn.microsoft.com/en-us/dotnet/api/system.io.bufferedstream). This class supports memory pooling and implements lazy buffer pattern
+
+<a href="https://www.nuget.org/packages/dotnext.net.cluster/5.17.0">DotNext.Net.Cluster 5.17.0</a>
+* Improved AOT support
+
+<a href="https://www.nuget.org/packages/dotnext.aspnetcore.cluster/5.17.0">DotNext.AspNetCore.Cluster 5.17.0</a>
+* Improved AOT support
+* Fixed [254](https://github.com/dotnet/dotNext/issues/254)
+
+<a href="https://www.nuget.org/packages/dotnext.maintenanceservices/0.5.0">DotNext.MaintenanceServices 0.5.0</a>
+* Improved AOT support
+
 # 12-07-2024
 <a href="https://www.nuget.org/packages/dotnext/5.16.0">DotNext 5.16.1</a>
 * Added [LEB128](https://en.wikipedia.org/wiki/LEB128) encoder and decoder as a public API. See `DotNext.Buffers.Binary.Leb128<T>` type for more information